refactor TR files
