+++ models/experimental/SSR/tt/__init__.py
@@ -7,14 +7,24 @@ from .tile_selection import (
     TTBasicLayer,
     TTMaskTokenInference,
 )
-from .patch_unembed import TTPatchUnEmbed
-from .window_attn_tr import TTWindowAttentionTR
+from .tile_refinement import (
+    TTWindowAttentionTR,
+    TTAttenBlocks,
+    TTHAB,
+    TTCAB,
+    TTWindowAttentionTR,
+    TTChannelAttention,
+    TTOCAB,
+    TTPatchUnEmbed,
+    TTPatchEmbed,
+)
 
 __all__ = [
     "TTMlp",
     "TTWindowAttention",
     "TTSwinTransformerBlock",
     "TTPatchEmbed",
+    "TTPatchUnEmbed",
     "TTPatchMerging",
     "TTBasicLayer",
     "TTMaskTokenInference",
